third_party/blink/web_tests/external/wpt/dom/nodes/querySelector-mixed-case-expected.txt

This is a testharness.js-based test.
[FAIL] Mixed HTML/SVG/MathML tree with various mixed-case attributes
  assert_equals: [viewbox] should only match HTML elements expected 2 but got 4
Harness: the test ran to completion.
